examples: remove outdated contact example files and consolidate functionality into a single example
